+++ src/main/java/com/kakaotech/team14backend/post/application/SchedulePostViewCount.java
@@ -1,6 +1,5 @@
-package com.kakaotech.team14backend.outer.post.schedule;
+package com.kakaotech.team14backend.post.application;
 
-import com.kakaotech.team14backend.inner.post.usecase.UpdatePostViewCountUsecase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -11,7 +10,7 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class SchedulePostViewCount {
 
-  private final UpdatePostViewCountUsecase updatePostViewCountUsecase;
+  private final UpdatePostViewCount updatePostViewCountUsecase;
 
   @Scheduled(initialDelayString = "${schedules.initialDelay}",fixedDelayString = "${schedules.fixedDelay}")
   public void execute() {
